RAG_01/pdf_RAG/requirements.txt
python-dotenv
langchain-community
langchain-text-splitters
pypdf
